+++ source/Programs/Compositor/HDR/lumStart.frag
@@ -54,7 +54,7 @@ mediump float Downscale4x4(sampler2D tex, const mediump vec2 uv, const mediump v
 
 //----------------------------------------------------------------------------------------------------------------------
 MAIN_PARAMETERS
-IN(highp vec2 vUV0, TEXCOORD0)
+IN(mediump vec2 vUV0, TEXCOORD0)
 
 MAIN_DECLARATION
 {
