+++ source/Programs/Compositor/raycast.vert
@@ -19,7 +19,7 @@ uniform highp mat4 WorldViewProj;
 MAIN_PARAMETERS
 IN(highp vec4 vertex, POSITION)
 IN(highp vec4 normal, NORMAL)
-OUT(highp vec2 vUV0, TEXCOORD0)
+OUT(mediump vec2 vUV0, TEXCOORD0)
 OUT(highp vec3 vRay, TEXCOORD1)
 
 MAIN_DECLARATION
